pyproject.toml
[project]
name = "lazyopenai"
version = "0.0.2"
description = ""
readme = "README.md"
authors = [{ name = "narumi", email = "toucans-cutouts0f@icloud.com" }]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.54.4",
    "pydantic-settings>=2.6.1",
]

[project.scripts]
lazyopenai = "lazyopenai.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.4",
    "types-pyyaml>=6.0.12.20240917",
]

[tool.ruff]
exclude = ["build"]
line-length = 120

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "C",   # flake8-comprehensions
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true
